+++ spec/data_forge/file/archiver_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe DataForge::File::Archiver do
+
+  describe ".from_input" do
+  end
+
+end